better epm rm

+++ cmd/epm/cli.go
@@ -10,6 +10,7 @@ import (
 	"github.com/eris-ltd/epm-go/epm"
 	"github.com/eris-ltd/epm-go/utils"
 	"github.com/eris-ltd/thelonious/monk"
+	"io/ioutil"
 	"log"
 	"os"
 	"path"
@@ -337,17 +338,28 @@ func cliRemove(c *cli.Context) {
 	if confirm("This will permanently delete the directory: " + root) {
 		// remove the directory
 		os.RemoveAll(root)
-		// remove from head (if current head)
-		_, h, _ := chains.GetHead()
-		if strings.Contains(root, h) {
-			chains.NullHead()
+		// we only remove refs if its not a multi
+		if !c.IsSet("multi") {
+			// remove from head (if current head)
+			_, h, _ := chains.GetHead()
+			if strings.Contains(root, h) {
+				chains.NullHead()
+			}
+			// remove refs
+			refs, err := chains.GetRefs()
+			ifExit(err)
+			for k, v := range refs {
+				if strings.Contains(root, v) {
+					os.Remove(path.Join(utils.Blockchains, "refs", k))
+				}
+			}
 		}
-		// remove refs
-		refs, err := chains.GetRefs()
-		ifExit(err)
-		for k, v := range refs {
-			if strings.Contains(root, v) {
-				os.Remove(path.Join(utils.Blockchains, "refs", k))
+		// if there are no chains left, wipe the dir
+		dir := path.Dir(root)
+		fs, _ := ioutil.ReadDir(dir)
+		if len(fs) == 0 {
+			if confirm("Remove the directory " + dir + "?") {
+				os.RemoveAll(dir)
 			}
 		}
 	}
